Fixing level 4 to new player setup

+++ Assets/Scripts/Collision/SpeedChangeCollisionAction.cs
@@ -27,7 +27,7 @@ namespace Collision
             Debug.Log("Player : " + playerMoveVector);
             
             
-            playerMoveVector.x = playerMoveVector.x * (float)1.3 * ((int)collisionEffectStrengthEnum / (int)CollisionEffectStrengthEnum.Normal);
+            playerMoveVector.x = playerMoveVector.x * (float)1.3 * ((float)collisionEffectStrengthEnum / (float)CollisionEffectStrengthEnum.Normal);
             playerMoveVector.y = playerMoveVector.y * (float)1.3 * ((int)collisionEffectStrengthEnum / (int)CollisionEffectStrengthEnum.Normal);
             playerMoveVector.z = playerMoveVector.z * (float)1.3 * ((int)collisionEffectStrengthEnum / (int)CollisionEffectStrengthEnum.Normal);
             Debug.Log("Player (SpeedChange) : " + playerMoveVector);
